🔥 Remove unused style codes and fix style
