+++ src/Page/QuestionDetailPage/Short/style.css.ts
@@ -1,92 +1,11 @@
-import {
-  createThemeContract,
-  createTheme,
-  style,
-  keyframes,
-  styleVariants,
-} from '@vanilla-extract/css';
+import { style, keyframes, styleVariants } from '@vanilla-extract/css';
 import { COLOR } from '../../../constants/color';
 import baseFontStyle from '../../../styles/font.css';
-
-export const vars = createThemeContract({
-  backgroundColor: null,
-  contentBackgroundColor: null,
-  textColor: null,
-});
-
-export const themeLightClass = createTheme(vars, {
-  backgroundColor: COLOR.WHITE,
-  contentBackgroundColor: COLOR.OFFWHITE,
-  textColor: COLOR.TITLEACTIVE,
-});
-
-export const themeDarkClass = createTheme(vars, {
-  backgroundColor: COLOR.TITLEACTIVE,
-  contentBackgroundColor: COLOR.DARKGRAY,
-  textColor: COLOR.WHITE,
-});
+import { vars } from '../baseStyle.css';
 
 const spread = keyframes({
   '0%': { backgroundColor: 'inherit' },
   '100%': { backgroundColor: vars.backgroundColor },
-});
-
-export const pageStyle = style({
-  boxSizing: 'border-box',
-
-  display: 'flex',
-  flexDirection: 'column',
-  flexWrap: 'wrap',
-  justifyContent: 'flex-start',
-  alignItems: 'stretch',
-  gap: '1.5rem',
-
-  width: '100%',
-
-  color: vars.textColor,
-  backgroundColor: vars.backgroundColor,
-
-  transition: 'background-color 0.2s',
-  animation: spread,
-  padding: '3rem',
-});
-
-export const topStyle = style({
-  display: 'flex',
-  alignItems: 'flex-start',
-  justifyContent: 'space-between',
-
-  width: '100%',
-});
-
-export const descStyle = style({
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'start',
-  justifyContent: 'center',
-});
-
-export const titleTagStyle = style({
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-  gap: '1.5rem',
-});
-
-export const questionContentStyle = style({
-  display: 'flex',
-  flexDirection: 'column',
-  gap: '2rem',
-  height: '60vh',
-  background: vars.contentBackgroundColor,
-  padding: '2rem',
-  borderRadius: '10px',
-});
-
-export const splitStyle = style({
-  display: 'flex',
-  width: '100%',
-  height: '100% ',
 });
 
 export const contentWrapperStyle = style({
@@ -128,23 +47,10 @@ export const answerInputTitleStyle = style([
 export const problemDescContentStyle = style([
   baseFontStyle.medium,
   {
-    overflowY: 'scroll',
+    overflow: 'auto',
     color: vars.textColor,
   },
 ]);
-
-export const questionDescStyle = style({
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'flex-start',
-  justifyContent: 'flex-start',
-  gap: '2.5rem',
-});
-
-export const answerInputStyle = style({
-  display: 'flex',
-  flexDirection: 'column',
-});
 
 const vibration = keyframes({
   from: { transform: 'rotate(1deg)' },
@@ -181,20 +87,11 @@ export const answerInputScoredStyle = styleVariants({
   default: [answerInputContentStyle, {}],
 });
 
-export const buttonListStyle = style({
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-  alignSelf: 'flex-end',
-  gap: '1rem',
-});
-
 export const tagListStyle = style({
   display: 'flex',
   gap: '0.25rem',
 });
 
-export const answerLengthButtonStyle = style({});
 export const answerLengthOpenStyle = style({
   visibility: 'visible',
 });
